requirements.txt
Flask
python-dotenv
google-generativeai
pytesseract
Pillow
PyMuPDF
Flask-Cors
requests
